refactor(redis): reorganize Redis queue implementation into multiple files
- Split queue.go into key.go, lease.go, message.go, option.go, and stream.go for better modularity
- Moved Redis key functions to key.go
- Extracted redisLease struct and Task method to lease.go
- Moved message handling and lease creation logic to message.go
- Migrated queue Option definitions and setters to option.go
- Relocated Redis stream operations and Lua promote script to stream.go
- Removed unused imports and constants from original file
- No functional changes, only code structure and organization improvements
